Code designed for use on Arduino controlled sumobots developed as part of UCD elecsoc
This code is designed to be implemented with an arduino powered sumobot.
Its aim is to hide away the more complex code required to use a variety
of motor drivers so as those new to arduino can focus on controlling the
sumobot itself. It should provide a higher level approach to coding
these types of robots, using a variety of possible combinations of
shields and sensors.